web: split lisa-html.ts (2326 -> 196 lines) — CSS + client JS into modules
The 2326-line single-file inlined GUI was flagged as near the maintainable
limit. Extract the <style> CSS to lisa-css.ts (MAIN_CSS) and the inline client
<script> to lisa-client.ts (MAIN_CLIENT_JS); lisa-html.ts recomposes them.

The template had zero ${} interpolations and zero inner backticks, so the
split is a verbatim slice — MAIN_HTML is byte-for-byte identical (len 76071,
sha256 unchanged), pinned by lisa-html-snapshot.test.ts so the GUI can't drift
silently. html-syntax.test.ts (vm.Script compile guard) still passes on the
recomposed script.
