+++ src/web/lisa-html-snapshot.test.ts
@@ -1,0 +1,30 @@
+import { test } from "node:test";
+import assert from "node:assert/strict";
+import { createHash } from "node:crypto";
+import { MAIN_HTML } from "./lisa-html.js";
+
+/**
+ * Byte-exactness guard for the lisa-html.ts split.
+ *
+ * MAIN_HTML used to be one ~2300-line template literal. It now stitches its
+ * <style> body and inline <script> body back together from lisa-css.ts
+ * (MAIN_CSS) and lisa-client.ts (MAIN_CLIENT_JS). Because no test can render
+ * the real GUI, the safety net for that refactor is "the served bytes did not
+ * change": identical output ⇒ identical browser behavior.
+ *
+ * These constants were captured from the pre-split MAIN_HTML. If you
+ * intentionally change the GUI markup/CSS/JS, recompute them:
+ *   node --import tsx -e 'import("./src/web/lisa-html.ts").then(m=>{const c=require("crypto");console.log(m.MAIN_HTML.length, c.createHash("sha256").update(m.MAIN_HTML).digest("hex"))})'
+ */
+const EXPECTED_LENGTH = 76071;
+const EXPECTED_SHA256 =
+  "77d2bc88b3bdb0bf0d595831dcc95118af456866d40ec1743966b92a80c80c17";
+
+test("MAIN_HTML length is byte-identical to the pre-split snapshot", () => {
+  assert.equal(MAIN_HTML.length, EXPECTED_LENGTH);
+});
+
+test("MAIN_HTML sha256 is byte-identical to the pre-split snapshot", () => {
+  const sha = createHash("sha256").update(MAIN_HTML).digest("hex");
+  assert.equal(sha, EXPECTED_SHA256);
+});
